system: make servod stop when devices get unplugged
As explained in [0]:

"It is important to realize how "TAG+="systemd", ENV{SYSTEMD_WANTS}=..." works.

The "systemd" tag tells systemd to instantiate a *.device unit to match the
device as seen by udev. The unit will be in the "active (plugged)" state as
long as the device is present.

"ENV{SYSTEMD_WANTS}=foo" tells systemd that the *.device unit should have
"Wants=" dependency on "foo". As the *.device unit is active, systemd will
fulfill this requirement dependency.

When the device goes away, it disappears from udev database and systemd
stops the *.device unit. Since the device is no more in the database,
there's nothing to attach the SYSTEMD_WANTS property to."

Therefore add the StopWhenUnneeded property to make systemd stop servod
once this happens and also add the KillMode=control-group option to
make systemd kill all spawned children of this servod instance.

With this change servo instances of individual boards get cleanup nicely
on unplug and of course also get reinstantiated after another plugin
without touching other servo instances on the same machine.

This also obsoletes the action="add" udev setting, as systemd only acts
on present devices anyway.

[0] https://bugzilla.redhat.com/show_bug.cgi?id=871074#c1

Signed-off-by: Heiko Stuebner <heiko@sntech.de>